backend/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
coverage==7.2.7
Django==4.1.13
django-cors-headers==3.13.0
django-livereload-server==0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==23.1
pathspec==0.11.1
pillow==10.3.0
platformdirs==3.5.1
PyJWT==2.8.0
python-decouple==3.7
pytz==2023.3
requests==2.32.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.5.0
tornado==6.3.3
urllib3==2.2.1
